+++ docs/changelog.md
@@ -37,6 +37,10 @@ Requires libvips v8.12.1
 * Expose `unlimited` option for SVG and PNG input, switches off safety features.
   [#2984](https://github.com/lovell/sharp/issues/2984)
 
+* Add `withoutReduction` option to resize operation.
+  [#3006](https://github.com/lovell/sharp/pull/3006)
+  [@christopherbradleybanks](https://github.com/christopherbradleybanks)
+
 ## v0.29 - *circle*
 
 Requires libvips v8.11.3